refactor: MainAppをapp.dartに切り出し
